+++ pcs/usage.py
@@ -768,6 +768,12 @@ _CLUSTER_CONFIG_SHOW_DESC = (
     "",
     _output_format_desc(),
 )
+_UTILIZATION_PLACEMENT_STRATEGY_DESC = (
+    """
+    For the utilization configuration to be in effect, cluster property
+    'placement-strategy' must be configured accordingly.
+    """,
+)
 
 
 def resource(args=()):
@@ -1222,6 +1228,8 @@ Commands:
         Utilization option should be in format name=value, value has to be
         integer. Options may be removed by setting an option without a value.
         Example: pcs resource utilization TestResource cpu= ram=20
+{utilization_placement_strategy_desc}
+
 
     relations <resource id> [--full]
         Display relations of a resource specified by its id with other resources
@@ -1407,6 +1415,9 @@ Notes:
             )
         ),
         update_desc=_format_desc(_resource_update_desc_fn(is_stonith=False)),
+        utilization_placement_strategy_desc=_format_desc(
+            _UTILIZATION_PLACEMENT_STRATEGY_DESC
+        ),
     )
     return sub_usage(args, output)
 
@@ -2851,7 +2862,12 @@ Commands:
         option should be in format name=value, value has to be integer.  Options
         may be removed by setting an option without a value.
         Example: pcs node utilization node1 cpu=4 ram=
-"""
+{utilization_placement_strategy_desc}
+""".format(
+        utilization_placement_strategy_desc=_format_desc(
+            _UTILIZATION_PLACEMENT_STRATEGY_DESC
+        )
+    )
     return sub_usage(args, output)