+++ pcs_test/tier1/legacy/test_node.py
@@ -10,6 +10,7 @@ from pcs import (
     utils,
 )
 
+from pcs_test.tier1.legacy.common import FIXTURE_UTILIZATION_WARNING
 from pcs_test.tools.assertions import AssertPcsMixin
 from pcs_test.tools.cib import get_assert_pcs_effect_mixin
 from pcs_test.tools.misc import get_test_resource as rc
@@ -88,7 +89,7 @@ class NodeUtilizationSet(
             self.temp_cib.name, "node utilization rh7-1 test1=10".split()
         )
         self.assertEqual(stdout, "")
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -102,7 +103,7 @@ class NodeUtilizationSet(
                 """
             ),
         )
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -117,7 +118,7 @@ class NodeUtilizationSet(
                 """
             ),
         )
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -125,7 +126,7 @@ class NodeUtilizationSet(
             "node utilization rh7-1 test1=-10 test4=1234".split(),
         )
         self.assertEqual(stdout, "")
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -140,7 +141,7 @@ class NodeUtilizationSet(
                 """
             ),
         )
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -148,7 +149,8 @@ class NodeUtilizationSet(
             "node utilization rh7-2 test2=321 empty=".split(),
         )
         self.assertEqual(stdout, "")
-        self.assertEqual(stderr, "")
+        self.assertEqual(retval, 0)
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -163,7 +165,7 @@ class NodeUtilizationSet(
                 """
             ),
         )
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -179,14 +181,14 @@ class NodeUtilizationSet(
                 """
             ),
         )
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
             self.temp_cib.name, "node utilization rh7-2 test1=-20".split()
         )
         self.assertEqual(stdout, "")
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -202,7 +204,7 @@ class NodeUtilizationSet(
                 """
             ),
         )
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
         stdout, stderr, retval = pcs(
@@ -218,48 +220,89 @@ class NodeUtilizationSet(
                 """
             ),
         )
-        self.assertEqual(stderr, "")
+        self.assertEqual(stderr, FIXTURE_UTILIZATION_WARNING)
         self.assertEqual(retval, 0)
 
     def test_refuse_non_option_attribute_parameter_among_options(self):
         self.assert_pcs_fail(
             "node utilization rh7-1 net".split(),
-            "Error: missing value of 'net' option\n",
+            (
+                f"{FIXTURE_UTILIZATION_WARNING}"
+                "Error: missing value of 'net' option\n"
+            ),
         )
 
     def test_refuse_option_without_key(self):
         self.assert_pcs_fail(
             "node utilization rh7-1 =1".split(),
-            "Error: missing key in '=1' option\n",
+            f"{FIXTURE_UTILIZATION_WARNING}Error: missing key in '=1' option\n",
         )
 
     def test_refuse_unknown_node(self):
         self.assert_pcs_fail(
             "node utilization rh7-0 test=10".split(),
-            "Error: Unable to find a node: rh7-0\n",
+            (
+                f"{FIXTURE_UTILIZATION_WARNING}"
+                "Error: Unable to find a node: rh7-0\n"
+            ),
         )
 
     def test_refuse_value_not_int(self):
         self.assert_pcs_fail(
             "node utilization rh7-1 test1=10 test=int".split(),
-            "Error: Value of utilization attribute must be integer: "
-            "'test=int'\n",
+            (
+                f"{FIXTURE_UTILIZATION_WARNING}"
+                "Error: Value of utilization attribute must be integer: "
+                "'test=int'\n"
+            ),
         )
 
     def test_keep_empty_nvset(self):
         self.assert_effect(
             "node utilization rh7-1 test=100".split(),
             self.fixture_xml_with_utilization(),
+            stderr_full=FIXTURE_UTILIZATION_WARNING,
         )
         self.assert_effect(
             "node utilization rh7-1 test=".split(),
             self.fixture_xml_empty_utilization(),
+            stderr_full=FIXTURE_UTILIZATION_WARNING,
         )
 
     def test_dont_create_nvset_on_removal(self):
         self.assert_effect(
             "node utilization rh7-1 test=".split(),
             self.fixture_xml_no_utilization(),
+            stderr_full=FIXTURE_UTILIZATION_WARNING,
+        )
+
+    def test_no_warning_printed_placement_strategy_is_set(self):
+        self.assert_effect(
+            "property set placement-strategy=minimal".split(),
+            self.fixture_xml_no_utilization(),
+        )
+        self.assert_resources_xml_in_cib(
+            """
+            <crm_config>
+                <cluster_property_set id="cib-bootstrap-options">
+                    <nvpair id="cib-bootstrap-options-placement-strategy"
+                        name="placement-strategy" value="minimal"
+                    />
+                </cluster_property_set>
+            </crm_config>
+            """,
+            get_cib_part_func=lambda cib: etree.tostring(
+                etree.parse(cib).findall(".//crm_config")[0],
+            ),
+        )
+        self.assert_effect(
+            "node utilization rh7-1 test=".split(),
+            self.fixture_xml_no_utilization(),
+        )
+        self.assert_effect(
+            "node utilization".split(),
+            self.fixture_xml_no_utilization(),
+            stdout_full="Node Utilization:\n",
         )
 
 
@@ -289,9 +332,10 @@ class NodeUtilizationPrint(TestCase, AssertPcsMixin):
     def test_refuse_when_node_not_in_mocked_cib(self):
         self.assert_pcs_fail(
             "node utilization some_nonexistent_node".split(),
-            [
-                "Error: Unable to find a node: some_nonexistent_node",
-            ],
+            (
+                f"{FIXTURE_UTILIZATION_WARNING}"
+                "Error: Unable to find a node: some_nonexistent_node\n"
+            ),
         )